Finished implementing fitbit record activity after powerlevel workout

+++ Powerlevel/Powerlevel/Powerlevel/App_Data/UP/USER_UP.sql
@@ -7,7 +7,7 @@
 CREATE TABLE [dbo].[User]
 (
 	[UserId] INT IDENTITY(1,1), 
-    [HeightFeet] FLOAT            NULL,
+    [HeightFeet] INT            NULL,
     [Weight] FLOAT NULL,
 	[DOB] DATETIME,
 	[Gender] NVARCHAR(10),
