+++ Powerlevel/Powerlevel/Powerlevel/Models/User.cs
@@ -19,7 +19,7 @@ namespace Powerlevel.Models
 
         public int UserId { get; set; }
 
-        public double? HeightFeet { get; set; }
+        public int? HeightFeet { get; set; }
 
         public double? Weight { get; set; }
